refactor(sentry): extract the run-record body builder from finalize.mjs
`scripts/sentry-autofix-finalize.mjs` sat at 1020 lines, over the
recurring-review checklist's 1000-line hard cap. Move the tracker run-record
rendering (`buildAutofixRunRecordBody` + `AUTOFIX_RUN_RECORD_MARKER` and their
private helpers) into a new `scripts/sentry-autofix-run-record.mjs`. Finalize
keeps the diff-guard / PR-body / label / marker-check decision layer and its
CLI, importing the builder back for the `run-record` subcommand — so the public
CLI entry points and subcommands are unchanged.

The pure builder tests move to `scripts/sentry-autofix-run-record.test.mjs`;
finalize's suite keeps the CLI `run-record` and rolling-comment-selection tests
and imports the marker from the new module. finalize.mjs is now 889 lines.

Wire the new Sentry suite in everywhere the #1721/#1779 coverage rules require:
a `sentry:autofix:run-record:test` package script pinned to its exact command,
a manifest entry in `scripts/sentry-suite-manifest.json` (floor 7), a direct
`node scripts/...` step in the ci.yml `scripts` job, the gate's trusted-alias
allowlist, and a quality-gate mapping that runs both the new suite and
finalize's.
