+++ .github/workflows/ci.yml
@@ -1160,6 +1160,11 @@ jobs:
         run: node scripts/sentry-autofix-select.test.mjs
       - name: Sentry autofix finalize suite
         run: node scripts/sentry-autofix-finalize.test.mjs
+      - name: Sentry autofix run-record suite
+        # The tracker run-record body builder, extracted from finalize.mjs when it
+        # crossed the 1,000-line cap; a scripts/sentry-*.test.mjs, so the #1721
+        # coverage rule requires a direct invocation here.
+        run: node scripts/sentry-autofix-run-record.test.mjs
       - name: Sentry triage archive suite
         run: node scripts/sentry-triage-archive.test.mjs
       - name: Sentry credential broker suite
