+++ package.json
@@ -53,6 +53,7 @@
     "sentry:autofix:select": "node scripts/sentry-autofix-select.mjs",
     "sentry:autofix:select:test": "node scripts/sentry-autofix-select.test.mjs",
     "sentry:autofix:finalize:test": "node scripts/sentry-autofix-finalize.test.mjs",
+    "sentry:autofix:run-record:test": "node scripts/sentry-autofix-run-record.test.mjs",
     "sentry:archive": "node scripts/sentry-triage-archive.mjs",
     "sentry:archive:test": "node scripts/sentry-triage-archive.test.mjs",
     "sentry:broker:test": "node --test scripts/sentry-mcp-broker.test.mjs",
